Add custom validation to User model
When a user is created through the Wagtail admin page, some fields can
interfere with others. Therefore, we have to prevent, for example, that
a user is being created as a customer and a coach at the same time.

This is done by defining error exceptions in the clean() method and then
calling full_clean() before saving the user object to call the
validating functions:
https://docs.djangoproject.com/en/2.2/ref/models/instances/#django.db.models.Model.full_clean

Also, some comments got adjusted.