Adjust some things in the User app
A user only gets set a password when it is initially created, not when
it is edited and already has set a password.

Skip the full_clean() validation on Superusers. They are for development
purposes only and therefore do not need to have set a phone number.

The string which is returned on calling the User object is changed to
return a string containing the phone number, first name, last name and
email address. This is used to identify the objects when selected e.g.
at Foreign Keys.