Add anamnese form page
A form page (AbstractEmailForm) is implemented. It consits of dynamic
form fields (AbstractFormField) which can be added in the Wagtail admin
CMS. The form can be submitted via GraphQL through our API.
The form submissions are saved as Anamnese objects.

Extra:
A date field is added to the anamnese model.
Remove the User import from anamnese, as it is not needed when using
Djangos get_user_model() function.
The Anamnese app is added to the GraphQL API in base settings file.
Also change some indents in the user models.py file.