Add beautyreport form page
A form page (AbstractEmailForm) is implemented. It consits of dynamic
form fields (AbstractFormField) which can be added in the Wagtail admin
CMS. The form can be submitted via GraphQL through our API.
The form submissions are saved as Beautyreport objects.

Extra:
A date field is added to the beautyreport model.
The Beautyreport app is added to the GraphQL API in base settings file.
Changed the used user model in the API. Will implement this in dynamic
API settings later.