Add anamnese GraphQL queries
The schema.py file in the anamnese app is added. A AnamneseType is
created which is based on the Anamnese model.
The schema contains four graphene types:
- List with all anamneses
- Field with one anamnese, filtered by the anamnese's id
- List with anamneses, filtered by a user id
- Field with the latest anamnese, filtered by by a user id

All queries can only be accessed with staff permissions.

The anamnese schema is then added to the API schema.py file.
Also some indents are changed here.

On saving a user, it is checked if the user is a coach, and if so, then
is_staff is also set to true.