Add beautyreport GraphQL queries
The schema.py file in the beautyreport app is added. A BeautyreportType
is created which is based on the Beautyreport model.

The schema contains six graphene types, following can be accessed only
with staff permissions:
- List with all beautyreports
- Field with one beautyreport, filtered by the beautyreport's id
- List with beautyreports, filtered by a user id
- Field with the latest beautyreport, filtered by by a user id
These can be accessed when logged in:
- List with beautyreports, filtered by the currently logged in user
- Field with the latest beautyreport, filtered by the current user

The beautyreport schema is then added to the API schema.py file.