Change User Identificator
The USERNAME_FIELD is changed to identificate users by its phone number.
In addition to the User Identificator change, an is_staff checkbox is
added to the User model and the phone number field has to be unique now.

The password and the is_staff field is added to the backend panels.

A coach does not have staff permissions anymore on creation.