Set correct Coach ForeignKey model
Previously, the ForeignKey for coaches in the user model was the user
model itself. Certainly, only objects saved as a coach should be an
option to be used as a ForeignKey.
This was more complicated than expected, as the user app and the coach
app do need to have circular imports for this.
I solved this with lazy loading:
https://stackoverflow.com/a/5680864/9638541