Rework phone login solution / Add queries to get a username
The AbstractUser Email field is added to the user model to set it to
be required before saving.
The USERNAME_FIELD variable was removed from the User model.
This variable was setting the main identificator to be the phone number,
but this was disturbing as it also applies to the wagtail login.

Because of that, two queries are added, by which the random generated
username of a user is given back when providing a valid phone number or
email address. This is available for anonymous users upwards.