Add extra Type for querying Coaches and Customers
The last commit changed the schema model for Coaches and Customers
which resulted in the User model's coach field not being able to be
queried anymore.
This got fixed by adding a class xxxQType which uses the base user model
which is used on graphene types, in addition to the CoachType class.