Adapt the Registration Form Page
The form page is renamed from FormPage to UserFormPage to identify it.
Fields are created which are needed to create a User. At the moment, the
form supports and needs the fields coach_id, first_name, last_name,
telephone and email. Coach_id gets referenced to a Coach object on
creation. A valid ID is offered by manually getting the latest given id
and adding 1 to it.