Add the remaining fields needed in the registration form
The optional form fields are added to be supported:
title, birthdate, address, city, postal_code, country and newsletter.
Birthdate is a Wagtail Date form field and gets automatically converted
to a Django DateField which is used in the model. Hereby, Validation is
already done on the first possible backend layer.
The country field accepts only two digits, otherwise a ValidationError
will be thrown by GraphQL.