Change the order and icons of the menu items
The order as well as the icons of the different menu tabs were not
created thoughtfully. Some are unnecessary displayed and some others had
the same menu icon.

The icons of some menu items are changed now and Model Groups are added:
The UserAdmin, AnamneseAdmin and BeautyreportAdmin classes are now
grouped together in a ModelAdminGroup within the Beautyreport app's
wagtail_hooks.py file.
Also the Coach and the Customer menu labels are renamed to be plural.