Add Coach field to Beautyreport and Anamnese Model
Add a field to the model of the Beautyreport and the Anamnese app.
The field 'coach' identifies from which coach account the anamnese or
beautyreport is submitted.

The fields are Foreign Keys with the Coach model.
The submitting coach is get with Wagtail's form.user function, which
returns the user object which submitted the form.
The new field is also shown in the overview panel and therefore added in
the wagtail_hooks.py file.

Also fix a typo in the Beautyreport GraphQL schema which did not let us
query the beautyreport lists by a specific user.