Add DOCX Generation Prototype
The first version of DOCX generation is added.
In the process_form_submission method of the anamnese app, user data
is manually added to be saved in the form_data variable. This makes it
easier to work with the user data on the beautyreport generation.

In the beautyreport app, the custom save function is added.
In this function, the data which is received through the submit
beautyreport form is processed. The received JSON string is translated
to a dictionary and a document is created based on this content.

Also the usage of datetime is changed, as before the python date was
used to get the current date. Now, best practice, Django's configured
timezone is used to get the current date and time.