Add BeautyreportDocument class
The class BeauyreportDocument, which exists in the Beautyreport app,
lists each link to a document w/ the corresponding beautyreport object.

A document field is also added to the Beautyreport model, which links to
the BeautyreportDocument model.

The template file is now placed within the media/-Folder.

When filling the document, the last for loop does not add a page break
anymore.

The document's name now depends on the BeautyreportDocument id, not the
Beautyreport id. Like this, no documents are overwritten.

The new class is added to wagtail hooks ModelAdmin and is named
"Beautyreport Links".