Fix ValueError when saving a Beautyreport
When creating a new beautyreport, a ValueError is thrown because I tried
to assign a Beautyreport ForeignKey to the BeautyreportDocument field
before the Beautyreport object is created.

To fix this, the super().Save method is called before creating the
BeautyreportDocument object.

The error message was:
save() prohibited to prevent data loss due to unsaved related object
'beautyreport'.