Add Set Password Mutation
The setPassword() mutation uses username, activation_token and password
as arguments. The activation_token argument must be the same as the
activation_url value in the specific user object for the mutation to be
executed.
The mutation returns a result boolean, which implies whether the mutation
was executed or not, a human-readable message and a single-string,
short message code.
short message code. At the moment, it fails on no_password_process or
on invalid_activation_token.

The mutation is added to the API schema.