Add PasswordResetMutation
The mutation to initialise the password reset process is added.
The mutation is called with passwodResetMutation(username).
When this mutation is called on a specific user name, the time of this
action is stored in a new field of the user model: last_password_reset.
Through this, the password reset process can only be started once in a
while (currently 900 seconds or 15 minutes).
A UUID is saved as the activation URL in the user object when
sucessfully executed.