Add Anamnese Excel Generation
Add a generation system like this for beautyreports but less complex.
The Excel generation uses the XlsxWriter module.
A anamnese is generated with the name "Anamnese-FIRST_NAME-DATE-ID" and
is then saved at the location defined in the base settings file at
AN_DOCUMENT_PATH, which is set to media/anamneses/.
A link to the document is saved as a AnamneseLink object.

This also adds the XlsxWriter module v1.2.3 to the base requirements.