Adapt Anamnese Excel Creation to alternative Form Data
Adapt the excel sheet generation to the new, alternative form data.
When creating an Anamnese object, form.full_values is now used instead
of form.cleaned_data.
Also a new format is added, which is applied to fields with numbers (align_left).

Finally, the fields which are shown per user are changed slightly.